runtime-local: extract the local task-drive layer from flowctl
`flowctl raw preview-next` and the forthcoming catalog-test runner need the same
thing — run a runtime-next task locally with synthetic shards, no Gazette broker,
no etcd, no Go consumer — and differ only in the three runtime-next host seams.

Moved out of crates/flowctl/src/raw/preview_next/: the three drivers, the
run-scoped services, the shard topology, and the segment-writing half of
fixture.rs (as segments.rs). flowctl is migrated onto the crate here, in the same
commit, so no divergent copy ever exists.

Two seams had to be generalized so the layer could stop knowing about preview:

- `Controls` is now generic over the publisher and logger factories.
- `start_with_shuffle_leader` takes a `build_shuffle` callback instead of a
  `fixture: bool`. This is what keeps journal *authorization* in flowctl: reading
  live journals needs a logged-in user token, and only flowctl has one.

The split of fixture.rs follows the same line: flowctl keeps the file format and
session planning, runtime-local takes the writer. Clock policy stays with the
caller — `write_transaction_for_bindings` accepts each document's clock — because
preview must keep reproducing the legacy harness's `3600 * ordinal + index`
stamping while a test runner paces clocks against its own synthetic time. That
routine also tags documents per *binding* rather than per collection, which the
catalog-test harness needs to honor per-transform read delays.

The synthetic topology is rebuilt on shared pieces: ranges from
`labels::shard::even_splits`, and shard IDs from `assemble::shard_id_prefix`
under generation zero, so a local run's shards take the shape they would carry in
production rather than `preview-derive/<task>/000`. Each driver likewise names
its runtime task after the task's own catalog name, which is what ops logs
attribute against.

`Controls` carries `--initial-state` and `--output-state` on shard zero's
`SessionLoop` rather than reaching into RocksDB by path, so no `Run` hosts a
RocksDB directory and each driver returns the final state its shard zero
reported at `Stopped`. `tests/connector_state.rs` covers that round trip.

Otherwise behavior-preserving, but for one defect the move made legible. `Run`
documented that its server task aborts on drop, but `_server_task` is a plain
`JoinHandle`, and dropping one *detaches* its task rather than aborting it — so
the tonic server outlived every `Run`. Add the `Drop` the comment assumed.
