+++ crates/runtime-local/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "runtime-local"
+version.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+authors.workspace = true
+readme = "./README.md"
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+publish = false
+
+[dependencies]
+assemble = { path = "../assemble" }
+doc = { path = "../doc" }
+json = { path = "../json" }
+labels = { path = "../labels" }
+models = { path = "../models" }
+ops = { path = "../ops" }
+proto-flow = { path = "../proto-flow" }
+proto-gazette = { path = "../proto-gazette" }
+runtime-next = { path = "../runtime-next" }
+service-kit = { path = "../service-kit" }
+shuffle = { path = "../shuffle" }
+
+anyhow = { workspace = true }
+bytes = { workspace = true }
+pbjson-types = { workspace = true }
+prost = { workspace = true }
+serde_json = { workspace = true }
+tempfile = { workspace = true }
+tokio = { workspace = true }
+tokio-stream = { workspace = true }
+tokio-util = { workspace = true }
+tonic = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+build = { path = "../build" }
+
+rustls = { workspace = true }
+tracing-subscriber = { workspace = true }
